backend/requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
SQLAlchemy==2.0.18
marshmallow==4.3.0
requests==2.31.0
gunicorn==21.2.0
